backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
python-dotenv
gTTS
openai-whisper
deepface
tf-keras